create comment_helper and move method to it
